pkg/columnstore: Add dynamic parquet
Dynamic parquet schemas in contrary to typical parquet schemas allow
for some variability in columns. In particular it automatically creates
column names dynamically when they are seen for the first time. This is
useful for data such as maps where the key may be known at search time,
but the value is high in cardinality. Using this scheme is good because
the columnar nature allows searching in those columns efficiently.